# packages/api-client/openapi/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
    "/api/auth/sign-up": {
      "post": {
        "operationId": "AuthController_signUp",
        "parameters": [],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Auth"
        ]
      }
    },
    "/api/auth/sign-in": {
      "post": {
        "operationId": "AuthController_signIn",
        "parameters": [],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Auth"
        ]
      }
    },
    "/api/auth/sign-out": {
      "post": {
        "operationId": "AuthController_signOut",
        "parameters": [],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Auth"
        ]
      }
    },
    "/api/auth/session": {
      "get": {
        "operationId": "AuthController_getSession",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Auth"
        ]
      }
    },
    "/api/risk/forecast": {
      "get": {
        "operationId": "RiskIntelligenceController_getForecast",
        "parameters": [
          {
            "name": "latitude",
            "required": true,
            "in": "query",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "longitude",
            "required": true,
            "in": "query",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "forecastDays",
            "required": false,
            "in": "query",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Get weather-driven risk forecast for a location",
        "tags": [
          "risk-intelligence"
        ]
      }
    },
    "/api/risk/user-impact": {
      "get": {
        "operationId": "RiskIntelligenceController_getUserImpact",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get impact summary for current user and family",
        "tags": [
          "risk-intelligence"
        ]
      }
    },
    "/api/families": {
      "post": {
        "operationId": "FamiliesController_createFamily",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateFamilyDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Create a family with permanent join code",
        "tags": [
          "families"
        ]
      }
    },
    "/api/families/join": {
      "post": {
        "operationId": "FamiliesController_joinFamily",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/JoinFamilyDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Join an existing family using code",
        "tags": [
          "families"
        ]
      }
    },
    "/api/families/me": {
      "get": {
        "operationId": "FamiliesController_getMe",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get authenticated user family memberships",
        "tags": [
          "families"
        ]
      }
    },
    "/api/families/me/map": {
      "get": {
        "operationId": "FamiliesController_getMyFamilyMap",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get map-focused family member visibility",
        "tags": [
          "families"
        ]
      }
    },
    "/api/volunteers/applications": {
      "post": {
        "operationId": "VolunteersController_apply",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/VolunteerApplicationDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Submit volunteer application",
        "tags": [
          "volunteers"
        ]
      }
    },
    "/api/volunteers/me/status": {
      "get": {
        "operationId": "VolunteersController_getStatus",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get volunteer profile/application status",
        "tags": [
          "volunteers"
        ]
      }
    },
    "/api/help-requests": {
      "post": {
        "operationId": "HelpRequestsController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateHelpRequestDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Create emergency help request",
        "tags": [
          "help-requests"
        ]
      }
    },
    "/api/help-requests/me": {
      "get": {
        "operationId": "HelpRequestsController_me",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "List my help requests and status history",
        "tags": [
          "help-requests"
        ]
      }
    },
    "/api/help-requests/{id}/claim": {
      "post": {
        "operationId": "HelpRequestsController_claim",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Claim help request as approved volunteer",
        "tags": [
          "help-requests"
        ]
      }
    },
    "/api/help-requests/{id}/status": {
      "patch": {
        "operationId": "HelpRequestsController_updateStatus",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateHelpRequestStatusDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Update claimed help request status",
        "tags": [
          "help-requests"
        ]
      }
    },
    "/api/warnings/me": {
      "get": {
        "operationId": "WarningsController_me",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get active warnings affecting current user",
        "tags": [
          "warnings"
        ]
      }
    },
    "/api/warnings/family": {
      "get": {
        "operationId": "WarningsController_family",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get active warnings affecting current family group",
        "tags": [
          "warnings"
        ]
      }
    },
    "/api/evacuation/areas": {
      "get": {
        "operationId": "EvacuationController_areas",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "List active evacuation areas",
        "tags": [
          "evacuation"
        ]
      }
    },
    "/api/evacuation/routes/suggested": {
      "get": {
        "operationId": "EvacuationController_suggested",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get warning-context route suggestions to shelters",
        "tags": [
          "evacuation"
        ]
      }
    },
    "/api/assistant/inquiries": {
      "post": {
        "operationId": "AssistantController_inquire",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AssistantInquiryDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Ask disaster preparedness assistant",
        "tags": [
          "assistant"
        ]
      }
    },
    "/api/admin/volunteers/applications/{id}/review": {
      "post": {
        "operationId": "AdminOperationsController_reviewVolunteer",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReviewVolunteerDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Approve or reject volunteer application",
        "tags": [
          "admin-operations"
        ]
      }
    },
    "/api/admin/warnings": {
      "post": {
        "operationId": "AdminOperationsController_createWarning",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateWarningDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Manually create and dispatch warning event",
        "tags": [
          "admin-operations"
        ]
      }
    },
    "/api/admin/vulnerable-regions": {
      "get": {
        "operationId": "AdminOperationsController_vulnerableRegions",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get vulnerable region dashboard data",
        "tags": [
          "admin-operations"
        ]
      }
    },
    "/api/admin/pins/statuses": {
      "get": {
        "operationId": "AdminOperationsController_pinStatuses",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get operational pin statuses",
        "tags": [
          "admin-operations"
        ]
      }
    },
    "/api/admin/warnings/prompt-suggestion": {
      "post": {
        "operationId": "AdminOperationsController_getPromptSuggestion",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/WarningPromptSuggestionDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get suggested warning prompt template",
        "tags": [
          "admin-operations"
        ]
      }
    }
  },
  "info": {
    "title": "Wira Borneo API",
    "description": "API documentation for local development and client generation",
    "version": "1.0.0",
    "contact": {}
  },
  "tags": [],
  "servers": [],
  "components": {
    "schemas": {
      "CreateFamilyDto": {
        "type": "object",
        "properties": {}
      },
      "JoinFamilyDto": {
        "type": "object",
        "properties": {}
      },
      "VolunteerApplicationDto": {
        "type": "object",
        "properties": {}
      },
      "CreateHelpRequestDto": {
        "type": "object",
        "properties": {}
      },
      "UpdateHelpRequestStatusDto": {
        "type": "object",
        "properties": {}
      },
      "AssistantInquiryDto": {
        "type": "object",
        "properties": {}
      },
      "ReviewVolunteerDto": {
        "type": "object",
        "properties": {}
      },
      "CreateWarningDto": {
        "type": "object",
        "properties": {}
      },
      "WarningPromptSuggestionDto": {
        "type": "object",
        "properties": {}
      }
    }
  }
}